Rearrange stats functions.
